backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
supabase==2.5.0
slowapi==0.1.9
python-multipart==0.0.9
